Refactor tool imports for consistency and clarity
