+++ sgr_deep_research/api/models.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Literal
 
 from pydantic import BaseModel, Field
 
-from sgr_deep_research.core.agents import (  # noqa: F403
+from sgr_deep_research.core.agents import (
     SGRAutoToolCallingResearchAgent,
     SGRResearchAgent,
     SGRSOToolCallingResearchAgent,
