+++ sgr_deep_research/core/agents/base_agent.py
@@ -14,9 +14,9 @@ from sgr_deep_research.core.models import AgentStatesEnum, ResearchContext
 from sgr_deep_research.core.prompts import PromptLoader
 from sgr_deep_research.core.stream import OpenAIStreamingGenerator
 from sgr_deep_research.core.tools import (
-    ClarificationTool,
     # Base
     BaseTool,
+    ClarificationTool,
     ReasoningTool,
     system_agent_tools,
 )
