+++ sgr_deep_research/core/tools/__init__.py
@@ -1,19 +1,23 @@
 from sgr_deep_research.core.tools.base import (
     AdaptPlanTool,
     AgentCompletionTool,
+    BaseTool,
     ClarificationTool,
     GeneratePlanTool,
     NextStepToolsBuilder,
     NextStepToolStub,
     ReasoningTool,
+    system_agent_tools,
 )
 from sgr_deep_research.core.tools.research import (
     CreateReportTool,
     WebSearchTool,
+    research_agent_tools,
 )
 
 __all__ = [
     # Tools
+    "BaseTool",
     "ClarificationTool",
     "GeneratePlanTool",
     "WebSearchTool",
@@ -23,4 +27,6 @@ __all__ = [
     "ReasoningTool",
     "NextStepToolStub",
     "NextStepToolsBuilder",
+    "system_agent_tools",
+    "research_agent_tools",
 ]
