+++ sgr_deep_research/settings.py
@@ -47,9 +47,7 @@ class PromptsConfig(BaseModel):
     """Prompts settings."""
 
     prompts_dir: str = Field(default="prompts", description="Directory with prompts")
-    tool_function_prompt_file: str = Field(
-        default="tool_function_prompt.txt", description="Tool functions prompt file"
-    )
+    tool_function_prompt_file: str = Field(default="tool_function_prompt.txt", description="Tool functions prompt file")
     system_prompt_file: str = Field(default="system_prompt.txt", description="System prompt file")
